+++ Rdmp.Core/Curation/Data/Aggregation/AggregateDimension.cs
@@ -176,7 +176,17 @@ namespace Rdmp.Core.Curation.Data.Aggregation
         /// <inheritdoc/>
         public string GetRuntimeName()
         {
-            return string.IsNullOrWhiteSpace(Alias) ? AggregateConfiguration.GetQuerySyntaxHelper().GetRuntimeName(SelectSQL):Alias;
+            if(string.IsNullOrWhiteSpace(Alias))
+            {
+                var syntax = _knownExtractionInformation.Value?.ColumnInfo?.GetQuerySyntaxHelper() ??
+                    AggregateConfiguration.GetQuerySyntaxHelper();
+
+                return syntax.GetRuntimeName(SelectSQL);
+            }
+            else
+            {
+                return Alias;
+            }
         }
 
         
@@ -228,6 +238,29 @@ namespace Rdmp.Core.Curation.Data.Aggregation
         public IHasDependencies[] GetObjectsDependingOnThis()
         {
             return new[] { AggregateConfiguration };
+        }
+
+        /// <summary>
+        /// Returns true if the <see cref="AggregateDimension"/> is likely to return a date based on the underlying column
+        /// data type.  If the <see cref="SelectSQL"/> is a transform this may be inaccurate.
+        /// </summary>
+        /// <returns></returns>
+        public bool IsDate()
+        {
+            var col = ColumnInfo;
+
+            if (col == null)
+                return false;
+
+            try
+            {
+                return col.GetQuerySyntaxHelper().TypeTranslater.GetCSharpTypeForSQLDBType(col.Data_type) == typeof(DateTime);
+            }
+            catch (Exception)
+            {
+                //it's some kind of wierd type eh?
+                return false;
+            }
         }
 
         public override void DeleteInDatabase()